GDGT-2299 make custom viz nav a bit smarter
