+++ frontend/src/metabase/admin/settings/components/SettingsNav/SettingsNav.tsx
@@ -1,20 +1,16 @@
-import { useDisclosure } from "@mantine/hooks";
-import React, { type ReactElement } from "react";
+import React from "react";
 import { t } from "ttag";
 
-import {
-  AdminNavItem,
-  type AdminNavItemProps,
-  AdminNavWrapper,
-} from "metabase/admin/components/AdminNav";
+import { AdminNavWrapper } from "metabase/admin/components/AdminNav";
 import { UpsellGem } from "metabase/common/components/upsells/components/UpsellGem";
 import { useHasTokenFeature, useSetting } from "metabase/common/hooks";
 import { PLUGIN_REMOTE_SYNC } from "metabase/plugins";
-import { getLocation } from "metabase/selectors/routing";
 import { getUserIsAdmin } from "metabase/selectors/user";
 import { Divider, Flex } from "metabase/ui";
 import { useSelector } from "metabase/utils/redux";
 
+import { CustomVisualizationsNav } from "./CustomVisualizationsNav";
+import { SettingsNavItem } from "./SettingsNavItem";
 import { UpdatesNavItem } from "./UpdatesNavItem";
 
 const NavDivider = () => <Divider my="sm" />;
@@ -28,10 +24,6 @@ export function SettingsNav() {
   const hasScim = useHasTokenFeature("scim");
   const hasPythonTransforms = useHasTokenFeature("transforms-python");
   const isHosted = useSetting("is-hosted?");
-  const hasCustomViz = useHasTokenFeature("custom-viz");
-  const customVizDevModeEnabled = useSetting(
-    "custom-viz-plugin-dev-mode-enabled",
-  );
   const isAdmin = useSelector(getUserIsAdmin);
 
   return (
@@ -69,36 +61,7 @@ export function SettingsNav() {
         icon="globe"
       />
       {/* do not allow users with "Settings access" permissions to access custom viz pages */}
-      {isAdmin && (
-        <SettingsNavItem
-          path={!hasCustomViz ? "custom-visualizations" : undefined}
-          folderPattern="custom-visualizations"
-          label={
-            <Flex gap="sm" align="center">
-              <span>{t`Custom visualizations`}</span>
-              {!hasCustomViz && <UpsellGem />}
-            </Flex>
-          }
-          icon="bar"
-        >
-          {hasCustomViz && [
-            <SettingsNavItem
-              key="manage"
-              path="custom-visualizations"
-              label={t`Manage visualizations`}
-            />,
-            ...(customVizDevModeEnabled
-              ? [
-                  <SettingsNavItem
-                    key="dev"
-                    path="custom-visualizations/development"
-                    label={t`Development`}
-                  />,
-                ]
-              : []),
-          ]}
-        </SettingsNavItem>
-      )}
+      {isAdmin && <CustomVisualizationsNav />}
       <SettingsNavItem path="maps" label={t`Maps`} icon="pinmap" />
       <SettingsNavItem
         path={!hasWhitelabel ? "whitelabel" : undefined}
@@ -153,77 +116,5 @@ export function SettingsNav() {
         icon="cloud"
       />
     </AdminNavWrapper>
-  );
-}
-
-/**
- * Find the child whose path is the best (longest) prefix match for the current
- * pathname, or null if no child matches at all. An exact match always wins.
- */
-const findBestMatchingChild = (
-  children: ReactElement[],
-  pathname: string,
-): ReactElement | null => {
-  let best: ReactElement | null = null;
-  let bestLen = 0;
-
-  for (const child of children) {
-    const childPath = child?.props?.path as string | undefined;
-    if (!childPath) {
-      continue;
-    }
-    const full = `/admin/settings/${childPath}`;
-    if (pathname === full || pathname.startsWith(`${full}/`)) {
-      if (full.length > bestLen) {
-        best = child;
-        bestLen = full.length;
-      }
-    }
-  }
-
-  return best;
-};
-
-export function SettingsNavItem({
-  path,
-  folderPattern,
-  active: activeOverride,
-  children: childrenProp,
-  ...navItemProps
-}: AdminNavItemProps & { active?: boolean }) {
-  const children = React.Children.toArray(childrenProp) as ReactElement[];
-  const currentPath: string = useSelector(getLocation)?.pathname ?? "";
-  const [isOpen, { toggle: toggleOpen }] = useDisclosure(
-    folderPattern ? currentPath.includes(folderPattern) : false,
-  );
-
-  const bestChild = findBestMatchingChild(children, currentPath);
-  const hasActiveDescendant = bestChild != null;
-
-  const fullPath = `/admin/settings/${path}`;
-  const showActive =
-    activeOverride ??
-    ((!isOpen && hasActiveDescendant) || currentPath === fullPath);
-
-  return (
-    <AdminNavItem
-      data-testid={`settings-sidebar-link`}
-      path={path ? `/admin/settings/${path}` : ""}
-      folderPattern={folderPattern}
-      opened={isOpen}
-      active={showActive}
-      onClick={toggleOpen}
-      {...navItemProps}
-    >
-      {children.length > 0
-        ? children.map((child) =>
-            child?.props?.path
-              ? React.cloneElement(child, {
-                  active: child === bestChild,
-                } as any)
-              : child,
-          )
-        : childrenProp}
-    </AdminNavItem>
   );
 }
