+++ frontend/src/metabase/admin/settings/components/SettingsNav/index.ts
@@ -1,1 +1,2 @@
 export * from "./SettingsNav";
+export * from "./SettingsNavItem";
